backend/requirements.txt
# LexAI — Backend Dependencies (Production)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
numpy>=1.24.0
pydantic>=2.0.0
gunicorn>=21.2.0

# File Processing
python-docx>=1.0.1
pdfplumber>=0.10.3
pytesseract>=0.3.10
pdf2image>=1.16.3
